Move soundboard into sound category, add usage to playclip
