+++ src/commands/soundCommands/playclip.js
@@ -1,0 +1,8 @@
+const { GenericSoundboardCommand } = require('../../models/')
+
+module.exports = new GenericSoundboardCommand({
+  triggers: ['playclip'],
+  description: 'Plays a custom sound clip',
+  usage: '{command} <clipname> <file attachment>',
+  existingConn: 'Develop the technology to speak two different things simultaneously and then we\'ll talk.'
+})
