--- src/commands/soundboardCommands/index.js
@@ -1,8 +1,0 @@
-const commands = require('fs').readdirSync(__dirname)
-  .filter(c => c !== 'index.js')
-  .map(c => require(`${__dirname}/${c}`))
-
-module.exports = {
-  commands,
-  name: '🎹 Soundboard Commands'
-}
